Added functionality for sorting locations - Added button in the tab bar that holds a menu picker to be able to select between name, date, and priority. - I seperated the content view into 2 views so I can leverage a more advanced query rather than hard coding a static sorting method.
