+++ RoamRadar/DestinationListingView.swift
@@ -1,0 +1,45 @@
+//
+//  DestinationListingView.swift
+//  RoamRadar
+//
+//  Created by Ethan Phillips on 10/6/23.
+//
+import SwiftData
+import SwiftUI
+
+struct DestinationListingView: View {
+    @Environment(\.modelContext) var modelContext
+    @Query(sort: [SortDescriptor(\Destination.priority, order: .reverse),
+                  SortDescriptor(\Destination.name)]) var destinations: [Destination]
+    
+    var body: some View {
+        List {
+            ForEach(destinations) { destination in
+                NavigationLink(value: destination) {
+                    VStack(alignment: .leading)  {
+                        Text(destination.name)
+                            .font(.headline)
+                        
+                        Text(destination.date.formatted(date: .long, time: .shortened))
+                    }
+                }
+            }
+            .onDelete(perform: deleteDestinations)
+        }
+    }
+    
+    init(sort: SortDescriptor<Destination>) {
+        _destinations = Query(sort: [sort])
+    }
+    
+    func deleteDestinations(_ indexSet: IndexSet) {
+        for index in indexSet {
+            let destination = destinations[index]
+            modelContext.delete(destination)
+        }
+    }
+}
+
+#Preview {
+    DestinationListingView(sort: SortDescriptor(\Destination.name))
+}
